login Req

+++ frontend/src/Pages/Electronics/Products.jsx
@@ -1,6 +1,7 @@
 import React, { useContext, useEffect, useState } from "react";
 import { UserContext } from "../Context/context";
 import URL from "../../server";
+import LoginPopUp from "../../components/LoginPopUp";
 const Products = ({
   id,
   name,
@@ -10,12 +11,16 @@ const Products = ({
   description,
   cartItems,
   setRefresh,
-  refresh
+  refresh,
+  setPopUP
 }) => {
   const { user } = useContext(UserContext);
-  console.log("cart",cartItems)
+  console.log("cart", cartItems);
   console.log("user", user);
   const removeFromCartHandler = async () => {
+  if(!user){
+    setPopUP(true)
+  }else{
     try {
       const response = await fetch(URL + "cart/" + user.id + "/" + id, {
         method: "DELETE",
@@ -29,34 +34,43 @@ const Products = ({
     } catch (err) {
       console.log("Error in add to Cart", err);
     }
+  }
   };
   const addToCartHandler = async () => {
-    try {
-      const response = await fetch(URL + "cart", {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({
-          id: user.id,
-          productId: id,
-          quantity: 1,
-        }),
-      });
-      if (response.ok) {
-        setRefresh(!refresh);
-        alert("added to cart");
-      } else {
-        alert("Failed to add");
+    if (!user) {
+      setPopUP(true)
+    } else {
+      try {
+        const response = await fetch(URL + "cart", {
+          method: "POST",
+          headers: {
+            "Content-Type": "application/json",
+          },
+          body: JSON.stringify({
+            id: user.id,
+            productId: id,
+            quantity: 1,
+          }),
+        });
+        if (response.ok) {
+          setRefresh(!refresh);
+          alert("added to cart");
+        } else {
+          alert("Failed to add");
+        }
+      } catch (err) {
+        console.log("Error in add to Cart", err);
       }
-    } catch (err) {
-      console.log("Error in add to Cart", err);
     }
   };
+  const buyHandler=()=>{
+    if(!user){
+      setPopUP(true)
+    }
+  }
   return (
     <div className="products-on-category-single">
       <div>
-        {" "}
         <img src={image} height="300px" alt={name} />
       </div>
       <div>
@@ -65,12 +79,12 @@ const Products = ({
         <p>{description}</p>
         <h6>Rs. {price}</h6>
         <div className="products-buttons">
-          {cartItems? (
+          {cartItems ? (
             <button onClick={removeFromCartHandler}>Remove from Cart</button>
           ) : (
             <button onClick={addToCartHandler}>Add to Cart</button>
           )}
-          <button>Buy</button>
+          <button onClick={buyHandler}>Buy</button>
         </div>
       </div>
     </div>
